Rename fancy.py to local.py amd include both the fancy and simplified display styles.
